cookbook/agents/nostr_rag/pyproject.toml
[project]
name = "nostr-rag-agent"
version = "0.1.0"
description = "A Nostr RAG agent for agentstr."
dependencies = [
    "python-dotenv",
    "langchain-openai",
    "agentstr @ {root:uri}/../../..",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
